Allow arrays for allow and deny in subnet pool
In the dhcp_subnet pool resource you can specify allow and deny
directives as strings, meaning you can only have one of each. Although
possible to achieve higher granularity with subclasses it is sometimes
simpler to just have two directives.

This commit allows for arrays or strings to be specified in the
resource, and coerces the string argument into a single-element array
for consistency.