Refactor LWRPs into custom resources
This originated from a need to fix the deprecation warnings for
use_inline_resources (FC113)[1] and this commit implements just the bare
minimum required to convert to custom resources while keeping all the
tests passing.

Support for chef-client 12 is dropped as that will be EOL this month[2].

I also removed the exec bit from some template and attribute files.

[1] http://www.foodcritic.io/#FC113
[2] https://www.chef.io/eol-chef12-and-chefdk1/